further modularization of code
