+++ src/components/listItem.ts
@@ -1,4 +1,8 @@
-export default function listItem() {
+export default function listItem(
+	convertSStoHHMMSS: (seconds: number) => string,
+	viewsFormatter: (views: number) => string,
+	unixTsFMT: (timestamp: number) => string
+) {
 	const listItemCSS = `
 :host {
   height: 20vmin;
@@ -87,40 +91,6 @@ display:block;
 }
 `;
 
-	function convertSStoHHMMSS(seconds: number): string {
-		const hh = Math.floor(seconds / 3600);
-		seconds %= 3600;
-		const mm = Math.floor(seconds / 60);
-		const ss = Math.floor(seconds % 60);
-		let mmStr = String(mm);
-		let ssStr = String(ss);
-		if (mm < 10) mmStr = '0' + mmStr;
-		if (ss < 10) ssStr = '0' + ssStr;
-		return hh > 0 ?
-			`${hh}:${mmStr}:${ssStr}` :
-			`${mmStr}:${ssStr}`;
-	}
-
-	const viewsFormatter = (views: number): string => Intl.NumberFormat('en', { notation: 'compact' }).format(views) + ' views';
-
-	function unixTsFMT(timestamp: number): string {
-		const seconds = (+new Date() - +new Date(timestamp)) / 1000;
-
-		const string =
-			seconds < 3600 ?
-				`${Math.floor(seconds / 60)} minute` :
-				seconds < 86400 ?
-					`${Math.floor(seconds / 3600)} hour` :
-					seconds < 604800 ?
-						`${Math.floor(seconds / 86400)} day` :
-						seconds < 2628000 ?
-							`${Math.floor(seconds / 604800)} week` :
-							seconds < 31536000 ?
-								`${Math.floor(seconds / 2628000)} month` :
-								`${Math.floor(seconds / 31536000)} year`;
-
-		return `${string}${string.startsWith('1 ') ? ' ' : 's'} ago`;
-	}
 
 
 	customElements.define('list-item', class extends HTMLElement {
